Move arg parser to own file
